requirements.txt
certifi==2019.11.28
cffi==1.13.2
cryptography==2.8
dnspython==1.16.0
emoji==0.5.4
future==0.18.2
pycparser==2.19
pymongo==3.10.0
python-telegram-bot==12.2.0
pytz==2019.3
six==1.13.0
tornado==6.0.3
